4.7.1 updates
