+++ integration_test/es_test.dart
@@ -3,8 +3,6 @@
  * See LICENSE for distribution and usage details.
  */
 
-import 'package:dotnet/utils/export.dart';
-
 import 'screens/home.dart' as home;
 import 'screens/settings_home.dart' as settings_home;
 import 'screens/products.dart' as products;
@@ -23,7 +21,7 @@ void main() async {
   WidgetsFlutterBinding.ensureInitialized();
 
   final Map<String, Object> testConfig = <String, Object>{
-    ...mobileDotnetConfig,
+    ...mobileEmpathConfig,
     appLocaleKey: <String>['es'],
   };
