+++ lib/screens/error.dart
@@ -40,7 +40,7 @@ class _ErrorScreenState extends State<ErrorScreen> {
   @override
   Widget build(BuildContext context) {
     return DotnetScaffold(
-      body: EzScreen(
+      EzScreen(
         Center(
           child: EzScrollView(
             mainAxisAlignment: MainAxisAlignment.center,
