--- lib/screens/products/export.dart
@@ -1,7 +1,0 @@
-/* dotnet
- * Copyright (c) 2025 Empathetech LLC. All rights reserved.
- * See LICENSE for distribution and usage details.
- */
-
-export 'for_creating.dart';
-export 'for_using.dart';
