+++ lib/screens/settings/color_settings.dart
@@ -14,10 +14,8 @@ class ColorSettingsScreen extends StatelessWidget {
   const ColorSettingsScreen({super.key, this.target});
 
   @override
-  Widget build(BuildContext context) => DotnetScaffold(
-        body: EzScreen(
-          EzColorSettings(target: target),
-          useImageDecoration: false,
-        ),
-      );
+  Widget build(BuildContext context) => DotnetScaffold(EzScreen(
+        EzColorSettings(target: target),
+        useImageDecoration: false,
+      ));
 }
