+++ pubspec.yaml
@@ -1,6 +1,6 @@
 name: dotnet
 description: "The Empathetech mission is to re-balance the power in tech by making self hosting more accessible."
-version: 4.6.3
+version: 4.7.1
 publish_to: 'none'
 
 environment:
@@ -21,7 +21,6 @@ dependencies:
   video_player: ^2.10.0
 
   # Community
-  after_layout: ^1.2.0
   feedback: ^3.2.0
   flutter_localized_locales: ^2.0.5
   flutter_platform_widgets: ^9.0.0
